ed25519 routines now support both Python 2 and 3. Moved duplicate code to one place in ed25519_utils.py.
